Changing module.exports to ES6 - Renaming Nav
