+++ components/index.js
@@ -1,11 +1,11 @@
-import Nav from './Nav'
+import Categories from './Categories'
 import Posts from './Posts'
 import Detail from './Detail'
 import Feedback from './Feedback'
 import Spinner from './Spinner'
 
 export {
-    Nav,
+    Categories,
     Posts,
     Detail,
     Feedback,
